requirements.txt
#cadquery
numpy
uharfbuzz
fpdf2
nicegui
PyMuPDF
tinycss